Python/pdfMerge/pyproject.toml
[tool.poetry]
name = "pdfmerge"
version = "0.1.0"
description = ""
authors = ["Michal Hloch <hloch.michal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypdf2 = "^3.0.1"
pymupdf = "^1.23.26"
aiohttp = "^3.9.3"
aiofiles = "^23.2.1"
pyyaml = "^6.0.1"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
